Finished V3 - Alphabeta Pruning

+++ src/ai/stats.rs
@@ -1,6 +1,6 @@
 pub struct SearchStatistics {
     pub num_position_evals: u32,
-    pub num_cutoffs: u64,
+    pub num_cutoffs: i32,
     pub think_time_ms: u32,
     pub num_checks: i32,
     pub num_mates: i32,
